Refactoring Weapon modal to Modal directory
